requirements.txt
aiohttp==3.10.10
langchain==0.3.3
langchain-chroma==0.1.4
langchain-cohere==0.3.0
langchain-community==0.3.2
langchain-core==0.3.10
langchain-experimental==0.3.2
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
mlflow==2.17.0
numpy==1.26.4
pandas==2.2.3
pydantic==2.9.2
python-dotenv==1.0.1
PyPDFLoader==0.1.0
streamlit==1.39.0
sentence-transformers==3.2.0
torch==2.4.1
transformers==4.45.2